pyproject.toml
[project]
name = "promptbin"
version = "0.1.0"
description = "Local-first prompt management and sharing tool for AI prompt engineers"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "markdown>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
